+++ packages/react-code-editor/src/components/CodeEditor/CodeEditor.tsx
@@ -16,10 +16,10 @@ import { TooltipPosition } from '@patternfly/react-core/dist/esm/components/Tool
 import { getResizeObserver } from '@patternfly/react-core/dist/esm/helpers/resizeObserver';
 import Editor, { BeforeMount, EditorProps, Monaco } from '@monaco-editor/react';
 import type { editor } from 'monaco-editor';
-import CopyIcon from '@patternfly/react-icons/dist/esm/icons/copy-icon';
+import RhUiCopyFillIcon from '@patternfly/react-icons/dist/esm/icons/rh-ui-copy-fill-icon';
 import UploadIcon from '@patternfly/react-icons/dist/esm/icons/upload-icon';
 import DownloadIcon from '@patternfly/react-icons/dist/esm/icons/download-icon';
-import CodeIcon from '@patternfly/react-icons/dist/esm/icons/code-icon';
+import RhUiCodeIcon from '@patternfly/react-icons/dist/esm/icons/rh-ui-code-icon';
 import HelpIcon from '@patternfly/react-icons/dist/esm/icons/help-icon';
 import Dropzone, { FileRejection } from 'react-dropzone';
 import { CodeEditorContext } from './CodeEditorUtils';
@@ -480,7 +480,12 @@ export const CodeEditor = ({
         const editorEmptyState =
           emptyState ||
           (isUploadEnabled ? (
-            <EmptyState variant={EmptyStateVariant.sm} titleText={emptyStateTitle} icon={CodeIcon} headingLevel="h4">
+            <EmptyState
+              variant={EmptyStateVariant.sm}
+              titleText={emptyStateTitle}
+              icon={RhUiCodeIcon}
+              headingLevel="h4"
+            >
               <EmptyStateBody>{emptyStateBody}</EmptyStateBody>
               {!isReadOnly && (
                 <EmptyStateFooter>
@@ -498,7 +503,12 @@ export const CodeEditor = ({
               )}
             </EmptyState>
           ) : (
-            <EmptyState variant={EmptyStateVariant.sm} titleText={emptyStateTitle} icon={CodeIcon} headingLevel="h4">
+            <EmptyState
+              variant={EmptyStateVariant.sm}
+              titleText={emptyStateTitle}
+              icon={RhUiCodeIcon}
+              headingLevel="h4"
+            >
               {!isReadOnly && (
                 <EmptyStateFooter>
                   <EmptyStateActions>
@@ -517,7 +527,7 @@ export const CodeEditor = ({
               <CodeEditorContext.Provider value={{ code: value }}>
                 {isCopyEnabled && (!showEmptyState || !!value) && (
                   <CodeEditorControl
-                    icon={<CopyIcon />}
+                    icon={<RhUiCopyFillIcon />}
                     aria-label={copyButtonAriaLabel}
                     tooltipProps={{
                       ...tooltipProps,
@@ -569,7 +579,7 @@ export const CodeEditor = ({
             {isLanguageLabelVisible && (
               <div className={css(styles.codeEditorTab)}>
                 <span className={css(styles.codeEditorTabIcon)}>
-                  <CodeIcon />
+                  <RhUiCodeIcon />
                 </span>
                 <span className={css(styles.codeEditorTabText)}>{language.toUpperCase()}</span>
               </div>