+++ packages/react-core/src/components/DualListSelector/examples/DualListSelectorComplexOptionsActions.tsx
@@ -21,13 +21,13 @@ import {
   MenuToggle,
   MenuToggleElement
 } from '@patternfly/react-core';
-import AngleDoubleLeftIcon from '@patternfly/react-icons/dist/esm/icons/angle-double-left-icon';
+import RhMicronsDoubleCaretLeftIcon from '@patternfly/react-icons/dist/esm/icons/rh-microns-double-caret-left-icon';
 import AngleLeftIcon from '@patternfly/react-icons/dist/esm/icons/angle-left-icon';
 import AngleDoubleRightIcon from '@patternfly/react-icons/dist/esm/icons/angle-double-right-icon';
 import AngleRightIcon from '@patternfly/react-icons/dist/esm/icons/angle-right-icon';
-import PficonSortCommonAscIcon from '@patternfly/react-icons/dist/esm/icons/pficon-sort-common-asc-icon';
+import RhMicronsSortDownSmallToLargeIcon from '@patternfly/react-icons/dist/esm/icons/rh-microns-sort-down-small-to-large-icon';
 import SearchIcon from '@patternfly/react-icons/dist/esm/icons/search-icon';
-import EllipsisVIcon from '@patternfly/react-icons/dist/esm/icons/ellipsis-v-icon';
+import RhUiEllipsisVerticalFillIcon from '@patternfly/react-icons/dist/esm/icons/rh-ui-ellipsis-vertical-fill-icon';
 
 interface Option {
   text: string;
@@ -155,7 +155,7 @@ export const DualListSelectorComplexOptionsActionsNext: React.FunctionComponent 
             aria-label="Sort Available"
             key="availableSortButton"
             isDisabled={isDisabled}
-            icon={<PficonSortCommonAscIcon />}
+            icon={<RhMicronsSortDownSmallToLargeIcon />}
           />,
           <Dropdown
             toggle={(toggleRef: React.Ref<MenuToggleElement>) => (
@@ -167,7 +167,7 @@ export const DualListSelectorComplexOptionsActionsNext: React.FunctionComponent 
                 variant="plain"
                 id="complex-available-toggle"
                 aria-label="Complex actions example available kebab toggle"
-                icon={<EllipsisVIcon />}
+                icon={<RhUiEllipsisVerticalFillIcon />}
               />
             )}
             isOpen={isAvailableKebabOpen}
@@ -190,7 +190,7 @@ export const DualListSelectorComplexOptionsActionsNext: React.FunctionComponent 
             aria-label="Sort Chosen"
             key="chosenSortButton"
             isDisabled={isDisabled}
-            icon={<PficonSortCommonAscIcon />}
+            icon={<RhMicronsSortDownSmallToLargeIcon />}
           />,
           <Dropdown
             toggle={(toggleRef: React.Ref<MenuToggleElement>) => (
@@ -202,7 +202,7 @@ export const DualListSelectorComplexOptionsActionsNext: React.FunctionComponent 
                 variant="plain"
                 id="complex-chosen-toggle"
                 aria-label="Complex actions example chosen kebab toggle"
-                icon={<EllipsisVIcon />}
+                icon={<RhUiEllipsisVerticalFillIcon />}
               />
             )}
             isOpen={isChosenKebabOpen}
@@ -283,7 +283,7 @@ export const DualListSelectorComplexOptionsActionsNext: React.FunctionComponent 
             isDisabled={chosenOptions.length === 0 || isDisabled}
             onClick={() => moveAll(false)}
             aria-label="Remove all"
-            icon={<AngleDoubleLeftIcon />}
+            icon={<RhMicronsDoubleCaretLeftIcon />}
           />
           <DualListSelectorControl
             onClick={() => moveSelected(false)}