+++ packages/react-core/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [6.6.0-prerelease.6](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-core@6.6.0-prerelease.5...@patternfly/react-core@6.6.0-prerelease.6) (2026-06-15)
+
+### Features
+
+- **Banner:** add pill variant ([#12440](https://github.com/patternfly/patternfly-react/issues/12440)) ([c91c911](https://github.com/patternfly/patternfly-react/commit/c91c911a0ee6d6329b4dec6135347efc2326040e))
+
+# [6.6.0-prerelease.5](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-core@6.6.0-prerelease.4...@patternfly/react-core@6.6.0-prerelease.5) (2026-06-11)
+
+**Note:** Version bump only for package @patternfly/react-core
+
+# [6.6.0-prerelease.4](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-core@6.6.0-prerelease.3...@patternfly/react-core@6.6.0-prerelease.4) (2026-06-11)
+
+**Note:** Version bump only for package @patternfly/react-core
+
+# [6.6.0-prerelease.3](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-core@6.6.0-prerelease.2...@patternfly/react-core@6.6.0-prerelease.3) (2026-06-11)
+
+**Note:** Version bump only for package @patternfly/react-core
+
+# [6.6.0-prerelease.2](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-core@6.6.0-prerelease.1...@patternfly/react-core@6.6.0-prerelease.2) (2026-06-10)
+
+**Note:** Version bump only for package @patternfly/react-core
+
 # [6.6.0-prerelease.1](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-core@6.5.1-prerelease.0...@patternfly/react-core@6.6.0-prerelease.1) (2026-05-26)
 
 ### Bug Fixes